Clean up 100% of static analyzer warnings and compile-time errors across the entire codebase workspace, silencing custom doc directives and deleting check_address.dart

+++ pkgs/num_dart/example/argsort_optimization_example.dart
@@ -1,5 +1,4 @@
 import 'package:num_dart/num_dart.dart';
-import 'dart:typed_data';
 
 void main() {
   print('=== NDArray Indirect Sorting (argsort) Optimization Examples ===\n');
@@ -22,10 +21,8 @@ void runArgsortSortingByKeysExample() {
   print('\nArgsort indices (sorted index mapping): ${indices.data}');
 
   // Map the names list according to the sorted indices!
-  final sortedNames = indices.data.map((idx) => names[idx as int]).toList();
-  final sortedPrices = indices.data
-      .map((idx) => prices.data[idx as int])
-      .toList();
+  final sortedNames = indices.data.map((idx) => names[idx]).toList();
+  final sortedPrices = indices.data.map((idx) => prices.data[idx]).toList();
 
   print('\n--- Sorted Results (By Price Ascending) ---');
   for (var i = 0; i < 4; i++) {
